refactor(mentor): base.js에서 constants.js로 파일명 변경
